Fix heap corruption with PHP 7.4 and uWSGI <2.0.19
Since bumping the PHP version to version 7.4 our tests started to fail,
flooding the logs with "zend_mm_heap corrupted" messages.

Using a newer version of uWSGI (2.0.19.1) didn't seem to have the issue,
so I bisected this and found that there indeed was a fix[1] regarding
this.

So while the upstream issue[2] doesn't seem to result in the same
behaviour as we're seeing in our test, the fix seems to be related
nevertheless.

To make sure we always have at least uWSGI version 2.0.19, I added a
conditional on the package, so even when using older versions of NixOS
(like we use in our tests) we still get a recent enough version of
uWSGI.

[1]: https://github.com/unbit/uwsgi/commit/06b6ea4cf74408963679fd63ed395
[2]: https://github.com/unbit/uwsgi/issues/2096

Signed-off-by: aszlig <aszlig@nix.build>