tests/upgrade: Re-add "epubreader" to tested apps
When updating Nextcloud apps in 989a136557cacfa56050dc60af87b3266c1477f,
I actually forgot about the assertion on epubreader 1.4.3, which was
exactly there so that we re-add epubreader to our tests.

I haven't run the full test yet, but according to some comment[1] in a
related issue, the issue has been fixed.

This fixes evaluation of our upgrade test.

[1]: https://github.com/e-alfred/epubreader/issues/18#issuecomment-720003796

Signed-off-by: aszlig <aszlig@nix.build>