lool: Patch out CapabilitiesService in app
The recent upgrade of richdocuments from version 3.7.6 to version 3.7.8
introduced a few more imports of CapabilitiesService, which we removed
in the first place since in our implementation, we already know the
capabilities and don't need to query them at runtime.

Unfortunately, I missed those new uses of CapabilitiesService when
rebasing the patch and we now get exceptions like this:

  Could not resolve OCA\Richdocuments\Capabilities! Class
  OCA\Richdocuments\Service\CapabilitiesService does not exist

I changed the patch to remove those references for now, but in the long
term we really need a proper integration test for libreoffice-online (or
well, Collabora Online for that matter).

Signed-off-by: aszlig <aszlig@nix.build>