updater: Prepare commit message with changelogs
So far my workflow was to run the updater and then copy & past all the
changelogs into the commit message editor.

While this works, it's still a bit inconvenient especially when I'm in
the middle of working on something else and now need to either use the
scrollback buffer or another terminal window.

This might not sound like a lot of distraction but if you're doing this
on a daily basis, it gets tedious.

Now we're creating commit message template with the changelog, so all we
need to do is stage the files and run "git commit ..." and we get all
the changes pre-filled in the commit message editor.

To do this, we're using the .git/SQUASH_MSG file, which is used during
an interactive rebase when you pick "squash". Ideally there would be
something like .git/PREPARED_MSG, since SQUASH_MSG is a bit confusing
but for our purposes it gets the job done :-)

Signed-off-by: aszlig <aszlig@nix.build>