Remove "diffuse" app
The app got removed from the Nextcloud app store and probably rightfully
so, since it looks like the app was never really doing anything and only
had an appinfo without any content.

Additionally, the upper bound of the version constraint was way to
high, so the app was also made available for Nextcloud 20 even though I
guess it wasn't really intended to be that way.

Here's the version constraint from appinfo/info.xml:

  <nextcloud max-version="99" min-version="10"/>

Looking at the old[1] app store entry, it seems that the app really
wasn't maintained at all, so that's probably why the author has finally
removed[2] the appinfo.

[1]: http://web.archive.org/web/20200814120316/https://apps.nextcloud.com/apps/diffuse
[2]: https://github.com/icidasset/diffuse/commit/46f3bba12270fcb7c677ec33bca7863ddee9a2f1

Signed-off-by: aszlig <aszlig@nix.build>