talk: Update signaling backend to version 0.2.0
The new version supports multiple backends now and thus our socket
activation patch needs a bit of rework. Our patch uses the new-style
configuration but still has a catch-all, which falls back to the
"default" backend because the only way to access the backend is via Unix
domain socket.

Alongside the signaling backend, I also updated Janus and its
dependencies to the latest upstream version.

This also marks the first upstream release version of
nextcloud-spreed-signaling, because so far we've used a pre-release
version and tracking the official releases is more likely that changes
in the Nextcloud app won't break communication with the signaling
backend.

Signed-off-by: aszlig <aszlig@nix.build>