talk: Fix admin settings with external signaling
Nextcloud Talk version 9.0.3 has introduced another call to
Config::getSignalingSecret, which in our case will fail since in our
case there is no secret and we exclusively talk with the backend via
Unix domain socket.

This fixes the following errors:

  Undefined index: secret at ...-nextcloud-20.0.2.2/apps/spreed/lib/Config.php
  Return value of OCA\\Talk\\Config::getSignalingSecret() must be of the
  type string, null returned

In the future it might be a good idea to introduce canaries so that
whenever an upstream change involves another call to the mentioned
function we get an error at *build* time rather than at runtime.

Signed-off-by: aszlig <aszlig@nix.build>