tests/upgrade: Exclude "permissions_overwrite"
When enabling the app, the error "Class 'Doctrine\DBAL\Types\Types' not
found" is thrown while applying the initial migrations.

I've reported this upstream at:

https://github.com/icewind1991/permissions_overwrite/issues/1

To make sure we re-enable the app, I added an assertion on the version
so we'll get an evaluation error for the next upstream release.

Signed-off-by: aszlig <aszlig@nix.build>