updater: Allow Nextcloud apps to be downgraded
The printer app is the first one since we have our updater which has
actually been downgraded and we didn't handle that. Essentially, since
our updater only recognises updates, I wouldn't have noticed, but
luckily the downgraded app is one where I opened an issue[1] about an
older Nextcloud version, so I was tracking what was going on.

What has happened is that the printer app got downgraded to version
0.0.3 for Nextcloud 18 and 19 while it got upgraded to version 0.0.5 for
Nextcloud 20.

To cope with this, we now just compare for version equality, rather than
greater/lesser and display apps as "downgraded" in the change summary
whenever the version went down.

[1]: https://github.com/e-alfred/nextcloud-printer/issues/13

Signed-off-by: aszlig <aszlig@nix.build>