tests/upgrade: Temporarily exclude "passwords" app
When upgrading Nextcloud 19 to Nextcloud 20, the migration for the
"passwords" app fails with the following error:

  You are using Nextcloud 19 but this release of the passwords app
  requires at least Nextcloud 20. See
  https://git.mdns.eu/nextcloud/passwords/wikis/Administrators/Migrations/Minimum-Requirement-Error
  for help.

This is probably not intended and the app should probably check the
version of the current source code rather than the *configured* version
of Nextcloud, the latter being only changed once the release upgrade is
finished.

I also added an assertion on the Nextcloud 20 version instead of the
Nextcloud 19 as done on most other (temporarily) excluded apps, because
the fix will most likely be included in the next release of the app for
Nextcloud 20.

The issue has already been reported upstream at:

https://github.com/marius-wieschollek/passwords/issues/330

Signed-off-by: aszlig <aszlig@nix.build>