Add Nextcloud major version 21
The biggest improvements Nextcloud Hub 21 introduces are:

  * New High Performance Back-end for Nextcloud Files: Reduces server
    load from desktop clients and web interface polling by 90% while
    delivering instant notifications to users
  * A wide range of performance improvements all over on top, decreasing
    loading times of pages and reducing load on the server
  * Collaborative features: new Whiteboard, author colours in Text and
    Document Templates
  * Nextcloud Talk: Debuts message status indicators, a raise hand
    feature, a group conversation description and more
  * A range of Groupware improvements like drag and drop and nicer
    threading in Mail and syncing social media avatars in Contacts

For more details, the announcement post can be found here:

https://nextcloud.com/blog/nextcloud-hub-21-out-with-up-to-10x-better-performance-whiteboard-and-more-collaboration-features/

On our side, we only needed to rebase the no-config-uid-check patch,
which didn't apply because a documentation URL in the code we're
removing has been fixed.

FIXME: Overhaul this commit (note to self: check *every* hunk and
especially remove all those forceEnable flags whenever possible) and
message as soon as there is an official release announcement.

TODO: Reword the paragraph below!

Disable documentserver_community, which is the binary package for
OnlyOffice. The main reason is to avoid hitting the output limit on
Hydra. However, we also want to avoid binary packages nevertheless,
since it's very unlikely that it will work on NixOS anyway.

Signed-off-by: aszlig <aszlig@nix.build>