Add Nextcloud major version 21
The biggest improvements Nextcloud Hub 21 introduces are:

  * New High Performance Back-end for Nextcloud Files: Reduces server
    load from desktop clients and web interface polling by 90% while
    delivering instant notifications to users
  * A wide range of performance improvements all over on top, decreasing
    loading times of pages and reducing load on the server
  * Collaborative features: new Whiteboard, author colours in Text and
    Document Templates
  * Nextcloud Talk: Debuts message status indicators, a raise hand
    feature, a group conversation description and more
  * A range of Groupware improvements like drag and drop and nicer
    threading in Mail and syncing social media avatars in Contacts

For more details, the announcement post can be found here:

https://nextcloud.com/blog/nextcloud-hub-21-out-with-up-to-10x-better-performance-whiteboard-and-more-collaboration-features/

On our side, we only needed to rebase the no-config-uid-check patch,
which didn't apply because a documentation URL in the code we're
removing has been fixed.

Since we no longer test Nextcloud 19 in our upgrade test, I re-enabled
gpoddersync because it only works with Nextcloud 20 and above.

I also re-enabled the talk_simple_poll app in the upgrade test, since
version 1.3.1 has been released a while ago and it is compatible with
Nextcloud 21.

The only app that's still causing issues (and thus is now excluded from
the upgrade test) is the "podcast" app, however the issue has already
been reported upstream so I'd expect it to be fixed soon.

Signed-off-by: aszlig <aszlig@nix.build>