tests/upgrade: Exclude "talk_simple_poll" app
The app fails with an exception when trying to enable it via the occ
command. This is probably because of NC 20/21 related changes that ended
up in the NC 19 version.

I also added an assertion as a reminder to re-add the app to the test
once a new upstream version is a available.

Signed-off-by: aszlig <aszlig@nix.build>
Issue: https://github.com/nickv-nextcloud/talk_simple_poll/issues/22