Consolidate extraConfig into nextcloud.settings
Since we already generate the whole configuration file via a Nix
function, we really don't need any "extraConfig" option, which for most
other services in NixOS would imply that it would be just a string to be
appended to some configuration file.

This is not the case however and instead it's merged with the attribute
set of the default configuration.

The downside is that it's not only more difficult to override options
but it also makes the configuration less introspectable by other
modules. For the push notification service however we need to have this.

While consolidating nextcloud.extraConfig into nextcloud.config would be
straightforward, I opted to use nextcloud.settings because it's more in
line with RFC-0042[1] and doesn't create an ambiguity with the "config"
module attribute.

[1]: https://github.com/NixOS/rfcs/blob/45b76f20add8d7e0e1d/rfcs/0042-config-option.md

Signed-off-by: aszlig <aszlig@nix.build>