tests/upgrade: Improve app list diff error message
So far the error message just listed all the provided apps and all the
expected apps in one inline string separated by "!=". This unfortunately
is a bit hard to read, so to make it more easy to spot the culprit we're
now using ndiff from difflib to make the exact differences more obvious.

Signed-off-by: aszlig <aszlig@nix.build>