tests/upgrade: Exclude more packages with binaries
All of them except "recognize" are just additional apps that should make
it easier for people to run various bundled binaries in an otherwise
PHP-only shared webhosting setting.

On NixOS however, these binaries do not work and we'd need to patchelf
them to get them to work. However, we do have source builds for those
packages that we need so we really need to unnecessarily blow up the
test closure size.

The "recognize" app bundles Node.js (yes, Node itself *and* various Node
modules) but instead of being just an add-on app for people being able
to get Node.js via some other means, the main functionality of the app
is also in the same tarball. However, since I currently dont' intent to
use this app, let's exclude it from the test as well.

To dump the sizes of the apps I used the following Nix expression:

  { pkgs ? import <nixpkgs> {}, lib ? pkgs.lib, majorVersion }:

  let
    path = ./packages + "/${toString majorVersion}/upstream.json";
    inherit (lib.importJSON path) applications;

    appSizes = lib.mapAttrsToList (name: attrs: let
      tarball = pkgs.fetchzip { inherit (attrs) url sha256; };
      nameEsc = lib.escapeShellArg name;
      tarballEsc = lib.escapeShellArg tarball;
      cmd = "echo \"$(du -sb ${tarballEsc} | cut -f1)\":${nameEsc}; ";
    in lib.optionalString (attrs ? url) cmd) applications;

  in pkgs.runCommand "sizes" {} ''
    (${lib.concatStrings appSizes}) | sort -t: -k1 -rn
  ''

Signed-off-by: aszlig <aszlig@nix.build>