Add a dedicated option to provide per-app patches
So far, all the patches were added via extraPostPatch to the main
Nextcloud derivation. This works fine so far but if you want to override
patches for a specific app from another module you can't do it because
you'd need to replace the whole extraPostPatch string.

Now eg. if you want to remove all patches for a particular app, all you
need to do is something like this:

  nextcloud.apps.spreed.patches = lib.mkForce [];

Another advantage of consolidating this is that we now get a message
that indicates which app is currently patched.

Signed-off-by: aszlig <aszlig@nix.build>