Add NC 20 compatibility patch for social app
Since this is the only app that holds us back to moving on to Nextcloud
21 and possibly 22, I decided to instead just apply the pull request[1]
that's been there since a few months and which fixes the Doctrine types
that causes the activation failure.

I usually try to avoid adding patches that are not related to something
specific to Nix(OS), but in this case I'll make an exception since it
holds us back from upgrading.

This reverts commit 7fcf6a91189527b98c7ed675671b67fd8f4df287, which
removed the app from our upgrade test.

[1]: https://github.com/nextcloud/social/pull/1249

Signed-off-by: aszlig <aszlig@nix.build>