tests/upgrade: Exclude pdfdraw app
This app needs a special server and additional configuration in the
reverse proxy. Since I'm not going to use this app and it's causing
failures, eg. when upgrading to Nextcloud 21, it's not worth the
trouble.

In case someone else wants the app, we can still re-enable it with the
proper support code.

Signed-off-by: aszlig <aszlig@nix.build>