Add patch to recognise Unix sockets in IpAddress
The IpAddress class is used for various things such as rate limiting and
throttling. Unfortunately whenever we connect via Unix domain sockets,
we'll run into an error like this:

  IpAddress.php:89: Trying to access array offset on value of type bool

In our case we use nginx as reverse proxy, which sets a REMOTE_ADDR of
"unix:". To cope with this, we're simply returing the address verbatim
when getSubnet() is called, since we don't have any concept of
subnetting with Unix domain sockets.

Signed-off-by: aszlig <aszlig@nix.build>