Add Nextcloud major version 22
The biggest improvements Nextcloud Hub 22 introduces are:

  * User-defined groups with Circles that makes it easier to manage
    teams where you can share files or assign tasks to circles, or
    create chat rooms for a circle
  * Integrated chat and task management where you can simply share a
    deck card into a chat room or turn a chat message into a task
  * Easy approval workflow, where an administrator can define a new
    approval flow in the settings and users can, on a document, request
    approval
  * Getting your document signatures easy with integrated PDF signing
    with DocuSign, EIDEasy, and LibreSign
  * Integrated knowledge management Nextcloud puts knowledge available
    to everyone at a moments’ notice, providing easy search, sharing,
    and portable access
  * Groupware improvements bringing a trash bin feature in Calendar and
    more. Nextcloud Mail features improved threading, email tagging, and
    support for Sieve filtering

For more details, the announcement post can be found here:

https://nextcloud.com/blog/nextcloud-hub-22-introduces-approval-workflows-integrated-knowledge-management-and-decentralized-group-administration/

On our side, we only needed to choose a longer password for our
Nextcloud Talk test, because now passwords need to be at least 10
characters or more.

Since the "Social" app is no longer maintained, I removed the app from
our tests.

Additionally, I added a patch for getting rid of the following warning:

  No application directories are marked as writable.

In our scenario *all* apps are residing in the Nix store and thus there
is no directory where we can write to.

For some reason, there is also a background job running that tries to
contact updates.nextcloud.com. I haven't seen this in earlier major
versions yet, so I'm not 100% sure why this suddenly is an issue.

Nevertheless, I added a patch that honors the "updatechecker" system
config option in the VersionCheck class.

Signed-off-by: aszlig <aszlig@nix.build>