module: Fix runner for Nextcloud cron
The Nextcroud cron runner didn't have our common PHP config applied, so
whenever there is eg. Redis involved, we'll get an error like this:

  Memcache \\OC\\Memcache\\Redis not available for local cache

Of course, we couldn't have used Redis before because the actual PHP
extension wasn't loaded because we were running the crons with a blank
PHP configuration.

Signed-off-by: aszlig <aszlig@nix.build>