apps: Fix tarball URLs with same content
The tarball for the "guests" app got moved into a special
"nextcloud-releases" organisation, but the original tarball location is
no longer available.

On the other hand, the files_fulltextsearch_tesseract app tarball points
to a fork which doesn't seem to have the generated tarballs (anymore?),
so I just changed the URLs to point to the source repository. While I
have no idea what happened over there, the contents are still the same
as the original tarball we had in our upstream.json files.

For both apps, the SHA-256 is still the same fortunately so all we need
to do here is switch to the new URL.

Signed-off-by: aszlig <aszlig@nix.build>