Update metadata of all apps
This is something we'd probably do automatically via our updater, but
when finding changed URLs and hashed I simply regenerated all
upstream.json files and so all the changed metadata has now been applied
on our side as well.

I haven't verified it but apparently it seems that it's possible to
change the metadata via the Nextcloud app store regardless of the
appinfo file.

Either way, the main reason I'm committing this is that we get rid of
the line noise for the next few changes which are not simply no-ops.

Signed-off-by: aszlig <aszlig@nix.build>