Remove "ocjobs" app for Nextcloud 19
The upload in question is bogus and the URL no longer exists upstream. I
also wasn't able to find a copy with the same contents and given that
the version in the tarball name doesn't match the app version I can only
guess that the upload was made in error and later deleted.

Neveretheless, the app in question is still published to the app
store[1] for Nextcloud 19 (and older) so to prevent errors on our side,
let's remove the app.

[1]: https://apps.nextcloud.com/apps/ocjobs

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Lea9250