Fix SHA-256 hashes for re-upladed app tarballs
While I personally don't like it whenever the upstream tarballs change
in contents and we really should go forwards to a future where all the
apps are built from source, the best we can do right now is simply
assume that the respective uploaders fumbled their release instead of
adding something malicious.

After all, that's the reason why app store tarballs are signed, so with
that in mind, let's assume the former.

For the change in the files_texteditor app it's a little more obvious
what happened, because the uploader probably copy & pasted the tarball
URL of the old version instead of the new release.

Signed-off-by: aszlig <aszlig@nix.build>