tests/upgrade: Exclude messagevault app
I guess that app has never been tested on Nextcloud 20, since the
initial migration already fails because the class 'OCP\DB\Types' can't
be found.

This is a common issue that we had with several apps already which
switched to the new[1] dbal types, but forgot about older Nextcloud
versions.

Since messagevault was never published with a version that has worked on
Nextcloud 20, I opened an upstream issue[1] suggesting to drop support
for Nextcloud 20 entirely.

Nevertheless, I added an assertion on the version on our side in the
unlikely event that they'll fix support for Nextcloud 20.

[1]: https://docs.nextcloud.com/server/latest/developer_manual/app_publishing_maintenance/upgrade-guide.html#updated-core-libraries
[2]: https://github.com/eblah/nextcloud-messagevault/issues/1

Signed-off-by: aszlig <aszlig@nix.build>