tests/upgrade: Exclude "vibeetoee"
I'm tired of chasing down all those weird apps that not only make it
hard to follow development but also don't have a dedicated bug report
URL.

Since first, the app is currently broken for Nextcloud 22, second, the
bug report URL is a corporate website and third, the upstream tarball is
using an unstable URL (https://mobikey.de/vibeetoee.tar.gz) I decided to
just exclude the app from our tests.

Nevertheless, I added an assertion on the version so that we at least
know whether things are improving. If that's not the case we can still
remove the assertion and permanently exclude the app.

Signed-off-by: aszlig <aszlig@nix.build>