updater: Don't include failed apps into changelist
Currently, we have a bunch of apps that fail to download and/or fail to
verify the signature, namely Grauphel, OC Jobs, Sendent and Sign Live.

Every time we do an update, we still get a changelist entry for them,
which is kinda confusing since we actually didn't update those apps.

To avoid this, let's just replace the new app release info with the
contents of the old app release info in the diff so that it no longer
gets included.

Signed-off-by: aszlig <aszlig@nix.build>