updater: Fix and improve types with latest nixpkgs
Newer nixpkgs also came with a newer Python and mypy version, which
improve type inferrence.

One particular thing that was caught by this was that the return types
of the ReleaseDiff.removed_apps and ReleaseDiff.added_apps properties
didn't actually make a lot of sense, so mypy complained with the
following error for both properties:

  error: Incompatible return value type (got "Dict[AppId, Optional[Any]]",
         expected "Dict[AppId, Union[InternalApp, ExternalApp]]")

For ReleaseInfo.removed_apps we don't actually need to return a dict,
because only the app IDs are used.

The ReleaseInfo.added_apps property created a dict with app IDs mapping
to None values, which is a bit odd because we actually have the
corresponding InternalApp/ExternalApp values available. We now create
that dict correctly.

Signed-off-by: aszlig <aszlig@nix.build>